+++ src/App.js
@@ -1,0 +1,7 @@
+import React from 'react';
+import Game from './components/Game'
+export default function App() {
+    return (
+        <Game />
+    )
+}
