+++ src/index.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
-import Game from './Game'
+import Game from './App'
 
 
 ReactDOM.render(<Game/>, document.getElementById("root"))
